Add function trigger retry policy (#11)
This PR adds an exponential backoff [retry strategy](https://learn.microsoft.com/en-us/azure/azure-functions/functions-bindings-error-pages?tabs=fixed-delay%2Cisolated-process%2Cnode-v4%2Cpython-v2&pivots=programming-language-python#retry-strategies) on the function. This strategy re-runs the function until it's successful, or retries are exhausted.
Also, since this function runs with an eventhub trigger, consumption from the partition that's feeding the function will not be restarted until after all the retries are exhausted.

The sizing of the policy is based in the following analysis, that the configuration
- min retry: 1s
- max retry: 7s
- max # of retries: 10

gives an expected wait time of 2m. This is what we've been using in Grafana's AWS Logs service loki.write components, but since this is a serverless runtime maybe this long-duration affects cost. Open to ideas here.

<img width="638" alt="image" src="https://github.com/user-attachments/assets/fad6c9c7-2346-4290-9727-c2bbcd65ef1d">

Also, I don't think this parameters should be parametrizable directly, rather, we could expose a more general setting like maximum wait time, and derive the (min, max, max retries) from that.

Closes https://github.com/grafana/grafana-csp-app/issues/224

Co-authored-by: Karsten Jeschkies <karsten.jeschkies@grafana.com>